Move supervisors to subdirectories
